fix(lint): satisfy golangci-lint (noinlineerr, unconvert)
- conf/imsconfig.go: split the inline `if err := DecodeHash(...); err != nil`
  into a plain assignment + check (noinlineerr).
- api/integration/default_password_test.go: drop the redundant int64() cast
  on created.PersonID, which is already int64 (unconvert).

+++ api/integration/default_password_test.go
@@ -104,7 +104,7 @@ func TestSetPersonPasswordDefault(t *testing.T) {
 	require.NoError(t, resp.Body.Close())
 
 	// ...is reset to the shared default.
-	resp = apisAdmin.setPersonPasswordDefault(ctx, int64(created.PersonID))
+	resp = apisAdmin.setPersonPasswordDefault(ctx, created.PersonID)
 	require.Equal(t, http.StatusNoContent, resp.StatusCode)
 	require.NoError(t, resp.Body.Close())
